# util/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API documentation for Leads app",
        "version": "1.0.0",
        "title": "Process Component API",
        "contact": {}
    },
    "basePath": "api/",
    "host": "localhost:3004/",
    "schemes": [
        "http"
    ],
    "paths": {
        "/leads": {
            "get": {
                "summary": "Finds all leads",
                "operationId": "listLeads",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid status value"
                    }
                }
            },
            "post": {
                "summary": "create Lead",
                "description": "give this link result as body https://run.mocky.io/v3/d026e567-47a8-4cee-b9ac-23d9123603ba",
                "operationId": "createLead",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/xml",
                    "application/json"
                ],
                "parameters": [
                
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Refer this link for body https://run.mocky.io/v3/d026e567-47a8-4cee-b9ac-23d9123603ba",
                        "required": true
                        
                      }
                ],
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                }
            }
        },
...
    },
...
}